[typeTag] Move TypeTag to the index file
Note: This forced some random other imports to change to prevent
import cycles
