+++ src/index.ts
@@ -7,6 +7,7 @@ export * from "./client";
 export * from "./core";
 export * from "./transactions/types";
 export * from "./transactions/typeTag";
+export * from "./transactions/typeTag/parser";
 export * from "./types";
 export * from "./utils/apiEndpoints";
 export * from "./utils/hdKey";
